refactor(solver): split query_cache.rs spread/cache accessors under 2000-line cap
Extract the trailing internal cache accessors and object-spread property
collection from the inherent impl<'a> QueryCache<'a> block into a new sibling
child module query_cache_spread.rs (declared via #[path] mod spread), mirroring
the existing query_cache_size.rs / query_cache_property.rs pattern. The moved
methods become pub(super) so the parent and sibling child modules retain
access; the Visibility import moves with its sole user. Pure code-move, no
behavior change. query_cache.rs drops from 2022 to 1787 lines.

AgentName: StudioOpus
